tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/csskeyframesrule_cssrules.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - cssRules on CSSKeyframesRule interface</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#dom-interfaces-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The cssRules attribute of CSSKeyframesRule
   is readonly and gives access to the keyframes in the list">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes test {
         from { width: 100px; }
         to { width: 200px; }
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed = false;
         try {
            var testKeyframes = document.styleSheets[0].cssRules[0];
            testKeyframes.cssRules = undefined;
            passed = (testKeyframes.cssRules.length == 2);
         } catch(e) {
         }
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS' : 'FAIL');
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
</body>
</html>
